move `getAllBadBitHashes`
